// NextMigration/lib/api/types.ts
export interface SetLike {
    id: string;
    user_id: string;
    set_id: string;
    set_type: 'studyset' | 'visualset';
    created_at: string;
}

export interface SetLikeListResponse {
    likes: SetLike[];
}
// ...
export interface CreateImage {
    title: string;
    url: string;
    index: number;
    grid_x: number;
    grid_y: number;
    scale: string;
}

export interface UpdateImage extends CreateImage {
    id: string;
}

export interface Image {
    id: string;
    title: string;
    url: string;
    index: number;
    grid_x: number;
    grid_y: number;
    scale: string;
    set_id: string;
}

export interface ImageWithPins extends Image {
    pins: PinListResponse;
}

export interface CreateVisualset {
    title: string;
    subject: string;
    folder_id: string;
    images: CreateImage[];
    pins: CreatePin[];
}

export interface UpdateVisualset {
    title?: string;
    public_set?: boolean;
    course?: string;
    folder_id?: string;
    images?: UpdateImage[];
    pins?: UpdatePin[];
}

export interface Visualset {
    id: string;
    title: string;
    course: string;
    created_at: string;
    last_updated: string;
    public_set: boolean;
    user_id: string;
    displayName: string;
    img_url: string;
    folder_id: string;
}

export interface FullVisualset extends Visualset {
    images: ImageWithPins[];
    likes: SetLikeListResponse;
    session?: StudySession;
    classrooms?: Classroom[];
}

export interface VisualsetListResponse {
    visualsets: Visualset[];
}

// ============================================================
// PINS (for Visualsets)
// ============================================================

export interface CreatePin {
    number: number;
    definition: string;
    x: number;
    y: number;
    image_id: string;
}

export interface UpdatePin extends Partial<CreatePin> {
    id: string;
}

export interface Pin {
    id: string;
    number: number;
    definition: string;
    x: number;
    y: number;
    created_at: string;
    updated_at: string;
    image_id: string;
    set_id: string;
    owner_id: string;
}

export interface PinListResponse {
    pins: Pin[];
}
// ...
export interface Classroom {
    id: string;
    name: string;
    owner_id: string;
    type: string;
    created_at: string;
    verified: boolean;
}
// ...
export interface StudySession {
    id: string;
    started_at: string;
    duration_min: number;
    ended_at?: string;
    set_index: number;
    accuracy: number;
    average_response_time: number;
    longest_focus_streak: number;
    last_seen?: string;
    last_studied?: string;
    user_id: string;
    set_id: string;
    set_type: 'studyset' | 'visualset';
}